[TS-348] Rename: 별자리 상세 관련 이름을 StarDetail로 변경
