+++ src/components/StarPage/StarDetail/Img.style.ts
