+++ src/components/StarPage/StarDetail/Img.tsx
@@ -1,15 +1,15 @@
 import starImg from "@/assets/image/img-card-detail-star.png";
 
-import { StarCardDetailStatus } from "@/types/star";
+import { StarDetailStatus } from "@/types/star";
 
 import { STAR_DETAIL_STATUS } from "@/constants/starPageConstants";
 
-import { imgStyle, labelStyle, starImgStyle } from "@/components/StarPage/StarCardDetail/Img.style";
+import { imgStyle, labelStyle, starImgStyle } from "@/components/StarPage/StarDetail/Img.style";
 
 const { PROGRESS, COMPLETE } = STAR_DETAIL_STATUS;
 
 interface ImgProps {
-	state: StarCardDetailStatus;
+	state: StarDetailStatus;
 	image: string;
 	starCount: number;
 }
