+++ src/components/StarPage/StarDetail/Story.style.ts
