+++ src/components/StarPage/StarDetail/Story.tsx
