+++ src/types/star.ts
@@ -1,6 +1,6 @@
 type TypeName = "육체활동" | "마음강화" | "지식습득" | "기록 습관" | "휴식 | 취미" | "";
 export type StarCardStatus = "PROGRESS" | "COMPLETE" | "NONE";
-export type StarCardDetailStatus = "SELECT" | "OTHER" | "PROGRESS" | "COMPLETE";
+export type StarDetailStatus = "SELECT" | "OTHER" | "PROGRESS" | "COMPLETE";
 export type CategoryType = { id: number; title: string; param: string };
 
 export interface StarDataType {
@@ -32,7 +32,7 @@ interface StarDetailType {
 	identity: string;
 	image: string;
 	starCount: number;
-	status: StarCardDetailStatus;
+	status: StarDetailStatus;
 	isProfile: boolean;
 }
